test(solid-query/useQueries): move type-only tests to 'useQueries.test-d.tsx'
